scx: lib - Split allocator declarations into sdt_alloc.h
The generic arena allocator machinery, the scx_alloc allocator, the static
and stack allocators, the buddy structs and scx_arena_subprog_init(), was
declared in sdt_task.h for historical reasons. Split it into sdt_alloc.h so
that non-task users do not have to include the task header. sdt_task.h
includes the new header, keeping every existing user unchanged.
